move legacy version to js-legacy
