+++ es6/README.md
@@ -25,7 +25,7 @@ jsMind 使用 ES6 编写，源代码位于 `/src`，构建目录为 `/es6`。
 
 ### 老版本下线
 
-之前的版本仍将会保留在 `/js` 目录里，我们将会在几个月后发布的新的版本里把这些文件移动到其它目录，并添加上不兼容性警告。
+之前的版本被移动到了 `/js-legacy` 目录里，我们将不再升级它，并会在未来的某个时间里删除此版本，建议所有人使用最新版本以获得最新的功能及bug修复。
 
 ### 本地构建
