+++ js/README.md
@@ -1,0 +1,8 @@
+ES6 版本现已发布，详情请查阅 [../es6/README.md](../es6/README.md) 。
+
+老版 jsmind 以及对应的插件已移至 `js-legacy` 目录下，其最终版本为 0.5.7 。我们将不再升级它，并会在未来的某个时间里删除它们，建议所有人使用最新版本以获得最新的功能及bug修复。
+
+
+The ES6 version of jsMind has been launched. see details from [../es6/README-en.md](../es6/README-en.md).
+
+Legacy jsmind.js and plugins have been moved to `/js-legacy`. we will no longer update them, and will remove them one day in the future. It's recommended that everyone use the latest version to follow continuous feature updates and bug fixes.
